# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Сервис Origummy",
        "description": "Сервис для магазина Origummy",
        "contact": {
            "email": "asekenskigo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/origummy"
        }
    ],
    "paths": {
...
        "/v1/categories/all": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Список всех категорий",
                "description": "Список всех категорий",
                "operationId": "getAllCategories",
                "responses": {
                    "200": {
                        "description": "Успешно получены"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/categories": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Список категорий",
                "description": "Список категорий",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Успешно получены"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "post": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Создать категорию",
                "description": "Создать категорию",
                "operationId": "createCategory",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория успешно создана"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/categories/search": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Поиск по названию категорий",
                "description": "Поиск по названию категорий",
                "operationId": "categorySearch",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Value"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
        "/v1/categories/{category}": {
            "get": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Получить категорию по ID",
                "description": "Получить категорию по ID",
                "operationId": "getCategoryById",
                "responses": {
                    "200": {
                        "description": "Данные по категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "put": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Изменить категорию",
                "description": "Изменить категорию",
                "operationId": "updateCategory",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория успешно изменена"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            },
            "delete": {
                "tags": [
                    "v1",
                    "admin",
                    "category"
                ],
                "summary": "Удалить категорию",
                "description": "Удалить категорию",
                "operationId": "deleteCategory",
                "responses": {
                    "200": {
                        "description": "Категория успешно удалена"
                    },
                    "400": {
                        "description": "Что-то не так"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryCreateRequest": {
                "required": [
                    "title",
                    "images"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "images": {
                        "description": "files:jpeg,png,jpg,svg",
                        "type": "array",
                        "items": {
                            "type": "image"
                        }
                    }
                },
                "type": "object"
            },
            "CategoryUpdateRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "images": {
                        "description": "files:jpeg,png,jpg,svg",
                        "type": "array",
                        "items": {
                            "type": "image"
                        }
                    },
                    "delete_images": {
                        "description": "file paths",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "CategoryResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "parent_id": {
                        "type": "integer"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "Object"
                        }
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "Object"
                        }
                    },
                    "parent": {
                        "type": "Object"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}